Criado o arquivo Conftest para disponibilizar as fixture entre os modulos
close #19
